Simple UI created; Created interfacebuilder.py and interfacegraph.py; Added tkinter and serial test
The UI contains only the visual design. It has no working
functions yet. It is build with building blocks from interfacebuilder.py.

The interfacebuilder.py contains easy building blocks for tkinter. This
allows me to stop worrying about tkinter anymore unless I need a different
type of widget. The widget are all in the same style so I can change a
single value to alter multiple widgets.

There is in main.py a class InterfaceVariables. This class as an object
holds the information for the whole interface. The information can be
found in any form such as in entries, labels, sliders or spinboxes.

build_interface() creates the tkinter application and calls the function
to generate the UI components. Those components are called panels for now
on. Each panel contains the defining widgets for the UI: entries, buttons,
information boxes, checkboxes, etc.

At last interfacegraph.py contains class Graph which produces a graph for
tkinter. Workings of this graph is not yet clear to me as I copied it from
a previous project. I verified in that project that the graph is working.