Program created to handle threading and serial communication
The program consist of manual test in code, thread class, and serial handler
class.

The serial handler class reads and writes data, connecting and disconnecting
the serial port, and formats and reorder the data from the reads. Reordering
data goes through a FIFO queue which is thread-safe.

The first test is about knowing what data the serial tools provides.

The second test runs a simple code, but it needs a serial connection to an
Arduino. The Arduino MUST send data to the program. AT the end the data gets
displayed to the terminal.

The last manual test creates a thread from the thread class. Then it starts
the thread and opens the serial port to the Arduino. It waits to collect for
the thread to collect the data and shows the first eight messages.

The test code checks if the thread class closes properly by using an Event
variable. Nothing special to see.

TODO:
At last interfacegraph.py contains class Graph which produces a graph for
tkinter. Workings of this graph is not yet clear to me as I copied it from
a previous project. I verified in that project that the graph is working.