Improved interfacegraph.py + minor changes to tests, interfacebuilder.py, and program.py
Added a new class GraphBase. This class should be an improvemnt to the
previous class Graph. GraphBase is able to show additional datasets without
doing any extra work from other functions (at least that is the whole idea).
Added manual test on how both graph classes needs to be programmed.

Renamed variables in the tests and program.py. program.py manual test has now
a less cluttered results.
Fixed bug SerialHandling trying to open a second serial connection.

DONE:
At last interfacegraph.py contains class Graph which produces a graph for
tkinter. Workings of this graph is not yet clear to me as I copied it from
a previous project. I verified in that project that the graph is working.