Improved the GUI with graph + expanded tests + minor improvements
The GUI has now functional buttons. The buttons are renamed to 'Connect',
'Disconnect', and 'Reconnect'. This has to do with the serial connection to
the Arduino. Improved the main loop with threading in mind.
After the connection is made the GUI will show the incoming data in a graph.
This graph slows down the GUI (which is bad and needs to be fixed).

The class GraphBase got an error handler (which is still very bad).

Moved class InterfaceVariables to interfacebuilder.py. It contains the
tkinter data, serial data (no idea what to name it), graph data (contains
data the graph can understand), arduino for serial tasks, and running flag.

Upgraded updatable label with an height scaler.
Added combobox with updatable selector.

Expanded tests:
TestSetupLibraries can be used to measure speed of the libraries. Changed
the asserts with success. It only fails when the library cannot be imported.
Added Numpy and Threading.
TestThreadingProgram now tests data of the serial format and queue.