Added new elements to the GUI + performance improvements
Added send entry to send data to the serial device. Made buttons for
data recording and saving. The buttons are skeletons for now.

Changed minor details to the interfacebuilder.py so that it intergrates
better with this GUI. The filters are now labeled with customizable names.

TODO: Make it that it comes from the settings file.

Class InterfaceVariables now manages all threads that are used. Starting
and closing threads should not be any worries anymore. Daemon threads can
be forced to close if closing is not working.

Threads are now labeled with names, so it is easier to spot which thread is
not responding to a close call. The update thread is split into two. The
first manages the data from the serial thread and transfers the data to the
graph. The second thread looks at the filters buttons and changes the
filter data to be equals to the GUI. The first thread handles the number
filtering.