Minor changes: made small GUI improvements and moved code aroud
Moved file panel to its own column. Moved filter panel to its own column.
This does not feel right at all but this improved the overview in the code.

The big thread functions are now seperated from the graph view panel
generation. This makes for the future possible to code unittest for
those functions.