Created feature saving data to CSV file in the GUI and background task
Start recording will make the program record the data it gets and pausing
stops the recording of the data. Saving can be done while recording or
better when paused. Data cannot be corrupted because list are thread safe.
Added message confirming that the user actions are completed.

The big thread functions are now split into smaller functions. This does
not make the code any better but it is cleaner to look at.
Created thread for handling data to csv file.

At last made the python script outside source so that the CSV files are
outside the source folder. TODO: add them to its own folder?
The files are made with filehandler.py. TODO: have the program remember
the previous settings of the last session.