SerialPlotter - Version 1.0
Fixed saved json data had no indents.
Fixed the settings were not saved randomly.
Fixed settings were not imported correctly to the GUI.

Reordered interface manager thread builder. The error message when exiting
the GUI still appears.

"Praparing program for large overhaul because implementing settings will
be very hard." - Yes it still will be because it is not done yet.
The graph filters are hard coded which is not the goal of being modular.